+++ core/impl/ModelLookup.js
@@ -4,47 +4,65 @@
  *
  * Copyright © 2019 Extremely Heavy Industries Inc.
  */
+import {forOwn} from 'lodash';
 import {createContext} from 'react';
-import {elemFactory} from '@xh/hoist/core';
+import {elemFactory, ModelPublishMode} from '@xh/hoist/core';
 
 /**
  * @private
  *
  * Support for making models available to components via React context.
- *
- * Not created directly by applications.  Components that need to *provide* models to
- * their descendants should use the WithModel component, which will create an instance
- * of this class.
+ * Not created directly by applications. Components specify how/if they publish/source their
+ * models from context via the `model` config option passed into the `hoistCmp()` factory.
+ * Hoist will in turn create and manage instances of this class to power those links.
  */
 export class ModelLookup {
     model;
     parent;
+    publishMode;
 
     /**
      * @param {Object} model -  model provided by this object.
-     * @param {ModelLookup} [parent] - parent instance of this class.
+     * @param {ModelLookup} parent - parent instance of this class.
+     * @param {ModelPublishMode} publishMode - mode for publishing this model to context.
      */
-    constructor(model, parent) {
+    constructor(model, parent, publishMode) {
         this.model = model;
         this.parent = parent;
+        this.publishMode = publishMode;
     }
 
     /**
-     * Lookup a model in the object, or one of its parents.
+     * Lookup a model in the context hierarchy
      *
-     * @param {(Class|string)} [selector] - class or name of mixin applied to class of
-     *      model to be returned.  If not provided the model contained in this object will be returned.
+     * @param {(Class|string)} selector - class or name of mixin applied to class of
+     *      model to be returned, or '*' to return the default model.
      * @returns {*} model or null if no matching model found.
      */
-    lookupModel(selector) {
-        const {model, parent} = this;
+    lookupModel(selector = '*') {
+        const {model, publishMode, parent} = this,
+            modeIsDefault = (publishMode === ModelPublishMode.DEFAULT),
+            isWildcard = (selector === '*');
 
-        if (!selector || selector == '*') return model;
+        // Try this model
+        if ((isWildcard && modeIsDefault) || (!isWildcard && model.matchesSelector(selector))) {
+            return model;
+        }
 
-        const ret = model.lookupModel(selector);
-        if (ret) return ret;
+        // Potentially try this model's direct children
+        if (modeIsDefault) {
+            let ret = null;
+            forOwn(model, (value, key) => {
+                if (value && value.isHoistModel && value.matchesSelector(selector)) {
+                    ret = value;
+                    return false;
+                }
+            });
+            if (ret) return ret;
+        }
 
-        return parent ? parent.lookupModel(selector) : null;
+        // Try parent
+        return parent?.lookupModel(selector) ?? null;
     }
 }