+++ desktop/cmp/panel/impl/Splitter.js
@@ -5,17 +5,18 @@
  * Copyright © 2019 Extremely Heavy Industries Inc.
  */
 
-import {hoistCmp} from '@xh/hoist/core';
+import {hoistCmp, uses} from '@xh/hoist/core';
 import {Icon} from '@xh/hoist/icon';
 import {button} from '@xh/hoist/desktop/cmp/button';
 import {hbox, vbox} from '@xh/hoist/cmp/layout';
+import {PanelModel} from '../PanelModel';
 
 import './Splitter.scss';
 
 
 export const splitter = hoistCmp.factory({
     displayName: 'Splitter',
-    model: false,
+    model: uses(PanelModel),
 
     render({model}) {
         const {vertical, collapsed, contentFirst, showSplitterCollapseButton, collapsible} = model,